Allow define the SBOM generation tool arch.
Adds one more input for the action used to install the SBOM generation
tool. The new input allow users to define the architecture to be used.
This is necessary to install the tool in multiple architecture like what
is done in the kwctl build process where the code is built for linux,
darwin and windows.