feat: run only lib tests for rust policies.
In the reusable workflow to perform basic validation in the Rust
policies, we run all the tests in the policy. This can cause issues if
the policy uses Rust to run e2e tests with kwctl. This commit updates
the workflow to run only lib tests which are the policy unit tests.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>