fix: ensure TinyGo policies build without errors
Latest version of Go introduces the concept of toolchain [1]. This
brings a new keyword inside of `go.mod`. This keyword can be understood
using Go 1.21 or later.

Even though our Go policies are built with TinyGo, TinyGo requires
the official Go compiler to perform some tasks.
Because of that, we have to ensure the latest stable release of the Go
compiler is installed inside of the build environment. We cannot rely on
what is shipped out of the box with the GH runner image.

[1] https://go.dev/blog/toolchain

Signed-off-by: Flavio Castelli <fcastelli@suse.com>