feat: Make policy-release action expect policy-version
By expecting as input of `policy-release` GHA, we push up the stack the
calculations of it.

policy-release `input.policy-version` will always be in the form of `0.1.0`.

Since now it's required, add it to all reusable workflows that call it.
No change on those workflows is needed, they were already calculating
their versions in their `calculate-version` steps.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>