fix: Always run calculate-version, correctly release policies
Now that policy-release.input.policy-version is required, always
calculate-version.

For the left side of the &&:
policy-release will push `:latest` for main runs, and `:v0.1.0` for
tags. Hence we don't need to check if the run was triggered by a tag.

For the right side of the &&:
We may want to push an artifact irrespectively if we want it published
on artifacthub, hence we don't need to check for inputs.artifacthub.

It may happen that for a run triggered without a tag, policy-version
will be empty or a previous existent tag, but it will not get used by
neither check-artifacthub nor policy-release anyways.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>