fix: Push fast-forward on artifacthub branch
When updating the artifacthub branch for releasing on artifacthub, do a
fast-forward with performing a `git pull` beforehand (with `git pull
--rebase` instead of plain `git pull` to not have merge commits).

If not, git will not accept the push as the remote contains work
that we do not. This is usually the case as we check out the repo
with actions/checkout@v4 and we don't check out every branch, hence
when we do a `git checkout artifacthub` we actually create a new
branch.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>