fix: temporarily downgrade tinygo version
Starting from tinygo 0.35.0, our policies cannot run because of an
issue inside of wapc's wasmtime-provider crate.

The issue has been fixed upstream and it will be part of the next release
of Kubewarden (1.21.0).

In the meantime, we have to downgrade the version of tinygo used by our GHA,
otherwise e2e tests will fail and it will be impossible to release fixes
for the tinygo policies.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>