feat: Add a new expect-sbom to policy-release action, and set it
Some languages like Rego don't have libraries, so we don't need to
create spdx SBOM files. Therefore, we shouldn't expect them when
creating the GH release, or the job for uploading the artifacts will
fail.

Hence, add a new `expect-sbom` input to the policy-release, separate
the upload of the spdx artifacts into a new step, and skip the step
when it isn't needed.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>
Co-authored-by: José Guilherme Vanz <jguilhermevanz@suse.com>