fix(renovate): group dependencies updates all together.
Changes the renovate bot configuration grouping all the dependencies
into a single PR to reduce the amount of PR to be reviewed. The
automerge has been disabled. Except to the lockfile maintenance PRs.

Furthermore, the renovate bot now is allowed to create the branches only
on weekends to avoid too much noise during the weekdays.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>